+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 ## v4.0.0 (2024-12-06)
 - Breaking changes
     - ⚠️ Drop support for PHP `7.1`, `7.2`, `7.3`, `7.4`, `8.0`, `8.1`
-    - ⚠️ Remove `UPCOMING` option that was deprecate in version `3`
-    - ⚠️ The `Option` class is now an enum with the following values: `ONLINE`, `NO_SUFFIX`, `JUST_NOW`
+    - ⚠️ The `Option` class is now an enum instead of a class
 - Improvements
     - 🧑‍💻 Codebase refactoring and cleanup
-    - ✨ Add support for PHP 8.4
+    - ✨ Added support for PHP 8.4
+    - ✨ Added new option `Option::RESET_CONF` to reset the configuration to default values
 
 ## [Release Notes v3](.github/CHANGELOG_V3.md)
 ## [Release Notes v2](.github/CHANGELOG_V2.md)